+++ ovp_users/serializers.py
@@ -1,10 +1,13 @@
 from django.core.exceptions import ValidationError
 from django.contrib.auth.password_validation import validate_password
+from django.contrib.auth import authenticate
 
 from ovp_users import models
+from ovp_uploads.serializers import UploadedImageSerializer
 
 from rest_framework import serializers
 from rest_framework import permissions
+from rest_framework import fields
 
 class UserCreateSerializer(serializers.ModelSerializer):
   class Meta:
@@ -28,23 +31,60 @@ class UserCreateSerializer(serializers.ModelSerializer):
     return super(UserCreateSerializer, self).validate(data)
 
 class UserUpdateSerializer(UserCreateSerializer):
+  current_password = fields.CharField(write_only=True)
+
   class Meta:
     model = models.User
     permission_classes = (permissions.IsAuthenticated,)
-    fields = ['name', 'phone', 'password', 'avatar']
+    fields = ['name', 'phone', 'password', 'avatar', 'current_password']
     extra_kwargs = {'password': {'write_only': True}}
 
+
+  def validate(self, data):
+    errors = dict()
+
+    if data.get('password') or data.get('current_password'):
+      current_password = data.pop('current_password', '')
+      password = data.get('password', '')
+
+      try:
+        validate_password(password=password)
+      except ValidationError as e:
+        errors['password'] = list(e.messages)
+
+      if not authenticate(email=self.context['request'].user.email, password=current_password):
+        errors['current_password'] = ["Invalid password."]
+
+    if errors:
+      raise serializers.ValidationError(errors)
+
+    return super(UserCreateSerializer, self).validate(data)
+
+
 class CurrentUserSerializer(serializers.ModelSerializer):
+  avatar = UploadedImageSerializer()
+
   class Meta:
     model = models.User
     fields = ['id', 'name', 'phone', 'avatar', 'email']
 
 class UserPublicRetrieveSerializer(serializers.ModelSerializer):
+  avatar = UploadedImageSerializer()
+
   class Meta:
     model = models.User
     fields = ['id', 'name', 'avatar']
 
+class UserProjectRetrieveSerializer(serializers.ModelSerializer):
+  avatar = UploadedImageSerializer()
+
+  class Meta:
+    model = models.User
+    fields = ['id', 'name', 'avatar', 'email', 'phone']
+
 class UserApplyRetrieveSerializer(serializers.ModelSerializer):
+  avatar = UploadedImageSerializer()
+
   class Meta:
     model = models.User
     fields = ['id', 'name', 'avatar', 'phone', 'email']